refactor(apps-backend): remove axios, change variables.
