+++ apps/apps-backend/package.json
@@ -17,7 +17,6 @@
         "@nestjs/core": "^10.0.0",
         "@nestjs/platform-express": "^10.0.0",
         "@nestjs/schedule": "^4.0.2",
-        "axios": "^1.7.2",
         "cache-manager": "^5.6.1",
         "reflect-metadata": "^0.2.2"
     },
