+++ apps/apps-backend/src/coingecko/coingecko.service.ts
@@ -3,10 +3,9 @@
 
 import { Inject, Injectable } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
-import axios from 'axios';
 import { CACHE_MANAGER, Cache } from '@nestjs/cache-manager';
 
-import { tokenPriceKey } from '../cache/cache.constants';
+import { TOKEN_PRICE_COINS, TOKEN_PRICE_CURRENCY, tokenPriceKey } from '../constants';
 
 @Injectable()
 export class CoinGeckoService {
@@ -15,25 +14,22 @@ export class CoinGeckoService {
     constructor(@Inject(CACHE_MANAGER) private cacheManager: Cache) {}
 
     async getTokenPrice(coinId: string, currency: string = 'usd'): Promise<number> {
-        try {
-            const response = await axios.get(`${this.baseUrl}/api/v3/simple/price`, {
-                params: {
-                    ids: coinId,
-                    vs_currencies: currency,
-                },
-            });
-            return response.data[coinId][currency];
-        } catch (error) {
+        const url = new URL(`${this.baseUrl}/api/v3/simple/price`);
+        url.searchParams.append('ids', coinId);
+        url.searchParams.append('vs_currencies', currency);
+
+        const response = await fetch(url.toString());
+        if (!response.ok) {
             throw new Error('Failed to fetch token prices from CoinGecko');
         }
+        const data = await response.json();
+        return data[coinId][currency];
     }
 
     @Cron(CronExpression.EVERY_HOUR)
     async refreshCachedPrices() {
-        const coins = ['iota'];
-        const currency = 'usd';
-        for (const coin of coins) {
-            const price = await this.getTokenPrice(coin, currency);
+        for (const coin of TOKEN_PRICE_COINS) {
+            const price = await this.getTokenPrice(coin, TOKEN_PRICE_CURRENCY);
             await this.cacheManager.set(tokenPriceKey(coin), price);
         }
     }
