+++ apps/apps-backend/src/constants.ts
@@ -2,3 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export const tokenPriceKey = (coinName: string) => `tokenPrice${coinName}`;
+export const TOKEN_PRICE_CURRENCY = 'usd';
+export const TOKEN_PRICE_COINS = ['iota'];
